refactor: change paths
