+++ src/components/cardsWrapper/cardsWrapper.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import CardsWrapper from './cardsWrapper';
-import itemsInfo from '../data/items';
+import itemsInfo from './../../data/items';
 
 describe('Cards wrapper', () => {
   it('renders component cards wrapper', () => {
