+++ src/components/searchBar/searchBar.css
@@ -46,7 +46,7 @@ input[type="search"]::-webkit-search-results-decoration {
   right: 10px;
   width: 30px;
   height: 30px;
-  background: url(./../assets/clear.svg) no-repeat center;
+  background: url(./../../assets/clear.svg) no-repeat center;
 }
 
 .search-button {
@@ -55,5 +55,5 @@ input[type="search"]::-webkit-search-results-decoration {
   right: 42px;
   width: 30px;
   height: 30px;
-  background: url(./../assets/search.svg) no-repeat center;
+  background: url(./../../assets/search.svg) no-repeat center;
 }
