+++ src/pages/AboutUsPage/AboutUsPage.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import AboutUsPage from './AboutUsPage/AboutUsPage';
+import AboutUsPage from './AboutUsPage';
 
 describe('About Us Page', () => {
   it('renders About Us page', () => {
